+++ backup_before_image_migration_20251201_192701.sql
@@ -1,0 +1,115 @@
+Initialising login role...
+Dumping schemas from remote database...
+17.6.1.054: Pulling from supabase/postgres
+dd242c929f22: Pulling fs layer
+1bdd3d43f094: Pulling fs layer
+6b1ebcb8bed9: Pulling fs layer
+d0c8f4f47ed8: Pulling fs layer
+ae233b33e2e4: Pulling fs layer
+93947ae0ebe2: Pulling fs layer
+97dd3f0ce510: Pulling fs layer
+4f4fb700ef54: Pulling fs layer
+0a3c806f1af6: Pulling fs layer
+77929950bfbd: Pulling fs layer
+4b70b07a812c: Pulling fs layer
+ef6b1cc6c775: Pulling fs layer
+4f4fb700ef54: Pulling fs layer
+9c90a5e2c7c0: Pulling fs layer
+c2ae11b175ef: Pulling fs layer
+5b9798e60c36: Pulling fs layer
+44e077b08f6b: Pulling fs layer
+236043153415: Pulling fs layer
+ee632b30d052: Pulling fs layer
+2dfa1d041c9b: Pulling fs layer
+c5907052ef04: Pulling fs layer
+8f5f5fd4e5d6: Pulling fs layer
+7c4069487cef: Pulling fs layer
+0bba81b1f1d9: Pulling fs layer
+126868546234: Pulling fs layer
+ec05ae2ace76: Pulling fs layer
+3a6b84a2c227: Pulling fs layer
+3192384ae596: Pulling fs layer
+c44eff0cab2a: Pulling fs layer
+9ad4dfe5f61c: Pulling fs layer
+9ad33fcfcadc: Pulling fs layer
+f6e7f25d89d8: Pulling fs layer
+6e23554173c1: Pulling fs layer
+95a3677a9b9a: Pulling fs layer
+a2ab0cdd8763: Pulling fs layer
+24d30e30e346: Pulling fs layer
+043f6dc52ce7: Pulling fs layer
+f492a419597e: Pulling fs layer
+c3c50defa27a: Pulling fs layer
+2ee63a9a5029: Pulling fs layer
+669d35859dd5: Pulling fs layer
+e514b425fb71: Pulling fs layer
+741ba396de2e: Pulling fs layer
+6013633eb90f: Pulling fs layer
+ddbc001ab68f: Pulling fs layer
+9564260991f0: Pulling fs layer
+895e5b881f37: Pulling fs layer
+d5fcf19145d1: Pulling fs layer
+4f4fb700ef54: Pulling fs layer
+9bff9d05965a: Pulling fs layer
+41d93667dfb9: Pulling fs layer
+e7f331d8540e: Pulling fs layer
+9731ea6b5f25: Pulling fs layer
+4f4fb700ef54: Already exists
+5b9798e60c36: Already exists
+6e23554173c1: Already exists
+669d35859dd5: Already exists
+9bff9d05965a: Already exists
+8f5f5fd4e5d6: Already exists
+24d30e30e346: Already exists
+6013633eb90f: Already exists
+9564260991f0: Already exists
+4b70b07a812c: Already exists
+c5907052ef04: Already exists
+41d93667dfb9: Already exists
+d0c8f4f47ed8: Already exists
+77929950bfbd: Already exists
+95a3677a9b9a: Already exists
+ef6b1cc6c775: Already exists
+e7f331d8540e: Already exists
+c2ae11b175ef: Already exists
+0a3c806f1af6: Already exists
+c44eff0cab2a: Already exists
+6b1ebcb8bed9: Already exists
+ee632b30d052: Already exists
+126868546234: Already exists
+a2ab0cdd8763: Already exists
+2ee63a9a5029: Already exists
+ddbc001ab68f: Already exists
+741ba396de2e: Already exists
+ec05ae2ace76: Already exists
+1bdd3d43f094: Already exists
+043f6dc52ce7: Already exists
+d5fcf19145d1: Already exists
+895e5b881f37: Already exists
+9731ea6b5f25: Already exists
+3a6b84a2c227: Already exists
+3192384ae596: Already exists
+c3c50defa27a: Already exists
+f492a419597e: Already exists
+e514b425fb71: Already exists
+236043153415: Already exists
+9c90a5e2c7c0: Already exists
+ae233b33e2e4: Already exists
+2dfa1d041c9b: Already exists
+9ad4dfe5f61c: Already exists
+9ad33fcfcadc: Already exists
+f6e7f25d89d8: Already exists
+93947ae0ebe2: Download complete
+97dd3f0ce510: Download complete
+97dd3f0ce510: Pull complete
+dd242c929f22: Download complete
+44e077b08f6b: Download complete
+3a6b84a2c227: Pull complete
+3192384ae596: Pull complete
+ae233b33e2e4: Pull complete
+44e077b08f6b: Pull complete
+0bba81b1f1d9: Download complete
+dd242c929f22: Pull complete
+4f4fb700ef54: Pull complete
+c44eff0cab2a: Pull complete
+7c4069487cef: Download complete